refactor : code cleanup - 경로 정리
