+++ frontend/src/App.tsx
@@ -1,19 +1,19 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import Header from '@/components/Header';
-import ErrorPage from '@/pages/ErrorPage/ErrorPage';
-import { AuthProvider } from '@/contexts/AuthContext';
+import Header from '@components/Header';
+import ErrorPage from '@pages/ErrorPage/ErrorPage';
+import { AuthProvider } from '@contexts/AuthContext';
 
 // Pages
-import HomePage from '@/pages/HomePage/HomePage';
-import SignInPage from '@/pages/SignInPage';
-import SignUpPage from '@/pages/SignUpPage';
-import WebtoonDetailPage from '@/pages/WebtoonDetailPage/WebtoonDetailPage';
-import MyPage from '@/pages/MyPage';
-import SocialLoginCallbackPage from '@/pages/SocialLoginCallbackPage/SocialLoginCallbackPage';
-import NewWebtoonsPage from '@/pages/NewWebtoonsPage';
-import OngoingWebtoonsPage from '@/pages/OngoingWebtoonsPage/OngoingWebtoonsPage';
-import CompletedWebtoonsPage from '@/pages/CompletedWebtoonsPage';
+import HomePage from '@pages/HomePage/HomePage';
+import SignInPage from '@pages/SignInPage';
+import SignUpPage from '@pages/SignUpPage';
+import WebtoonDetailPage from '@pages/WebtoonDetailPage/WebtoonDetailPage';
+import MyPage from '@pages/MyPage';
+import SocialLoginCallbackPage from '@pages/SocialLoginCallbackPage/SocialLoginCallbackPage';
+import NewWebtoonsPage from '@pages/NewWebtoonsPage';
+import OngoingWebtoonsPage from '@pages/OngoingWebtoonsPage/OngoingWebtoonsPage';
+import CompletedWebtoonsPage from '@pages/CompletedWebtoonsPage';
 
 const App: React.FC = () => {
   return (
