+++ frontend/src/models/props.ts
